+++ src/Commands/codebin/sourcebin.js
@@ -1,5 +1,5 @@
 const { Command } = require('discord.js-commando');
-const { create } = require('sourcebin');
+const { create, url } = require('sourcebin');
 
 module.exports = class SourcebinCommand extends Command {
 	constructor(client) {
@@ -19,11 +19,12 @@ module.exports = class SourcebinCommand extends Command {
 	}
 
 	async run(msg, { code }) {
-		const url = create([{ content: code.code, language: 'text' }], {
+		const key = await create([{ content: code.code, language: 'text' }], {
 			title: 'CodeBin',
 			description: 'Uploaded by TanakaBot <https://github.com/1chiSensei/Tanaka>',
 		});
+		const link = url(key);
 
-		return msg.say(`The link to the code is: ${url}!`);
+		return msg.say(`The link to the code is: ${link}!`);
 	}
 };